Pvp request async v2
